Fixed failing migration

+++ db/migrate/20120305172415_remove_unused_sousreseaux_table.rb
@@ -1,6 +1,6 @@
 class RemoveUnusedSousreseauxTable < ActiveRecord::Migration
   def up
-    drop_table :sousreseaux
+    drop_table :sousreseaux if table_exists?(:sousreseaux)
   end
 
   def down
